refactor: tweak button style code (#51646)
